Add constraint to shared algo for cleaner access
