+++ crates/fuel-core/src/service/adapters/fuel_gas_price_provider/tests/producer_gas_price_tests.rs
@@ -1,11 +1,11 @@
 use crate::service::adapters::fuel_gas_price_provider::{
-    ports::GasPriceAlgorithm,
     tests::{
         build_provider,
         TestGasPriceAlgorithm,
     },
 };
 use fuel_core_producer::block_producer::gas_price::GasPriceProvider as ProducerGasPriceProvider;
+use fuel_core_gas_price_service::GasPriceAlgorithm;
 
 #[tokio::test]
 async fn gas_price__if_requested_block_height_is_latest_return_gas_price() {
@@ -15,7 +15,7 @@ async fn gas_price__if_requested_block_height_is_latest_return_gas_price() {
     let bytes = 10;
 
     // when
-    let expected_price = algo.gas_price(bytes);
+    let expected_price = algo.next_gas_price(bytes);
     let actual_price = gas_price_provider.gas_price(bytes).await.unwrap();
 
     // then
