+++ crates/fuel-core/src/service/adapters/fuel_gas_price_provider/tests/tx_pool_gas_price_tests.rs
@@ -1,11 +1,11 @@
 use crate::service::adapters::fuel_gas_price_provider::{
-    ports::GasPriceAlgorithm,
     tests::{
         build_provider,
         TestGasPriceAlgorithm,
     },
 };
 use fuel_core_txpool::ports::GasPriceProvider as TxPoolGasPriceProvider;
+use fuel_core_gas_price_service::GasPriceAlgorithm;
 
 #[tokio::test]
 async fn gas_price__if_requested_block_height_is_latest_return_gas_price() {
