+++ crates/services/gas_price_service/src/static_updater.rs
@@ -1,4 +1,4 @@
-use crate::UpdateAlgorithm;
+use crate::{GasPriceAlgorithm, UpdateAlgorithm};
 use async_trait::async_trait;
 use fuel_core_types::fuel_types::BlockHeight;
 
@@ -26,6 +26,19 @@ impl StaticAlgorithm {
     }
 }
 
+impl GasPriceAlgorithm for StaticAlgorithm {
+    fn last_gas_price(&self) -> u64 {
+        self.price()
+    }
+
+    fn next_gas_price(&self, _block_bytes: u64) -> u64 {
+        self.price()
+    }
+
+    fn worst_case_gas_price(&self, _block_height: BlockHeight) -> u64 {
+        self.price()
+    }
+}
 #[async_trait]
 impl UpdateAlgorithm for StaticAlgorithmUpdater {
     type Algorithm = StaticAlgorithm;
